NewsUPD

+++ serviceachat/achat/views.py
@@ -2,8 +2,35 @@ from http import client
 from django.shortcuts import render
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+#from django.contrib.auth import authenticate
+#from django.views.decorators.csrf import csrf_exempt
+#from rest_framework.authtoken.models import Token
+#from rest_framework.decorators import api_view, permission_classes
+#from rest_framework.permissions import AllowAny
+#from rest_framework.status import (
+#    HTTP_400_BAD_REQUEST,
+#    HTTP_404_NOT_FOUND,
+#    HTTP_200_OK
+#)
 from .serializers import clientSerializer, articleSerializer, panierSerializer, lignAchatSerializer, LigneAchat
 from .models import Article, Client, LigneAchat, Panier
+
+#@csrf_exempt
+#@api_view(["POST"])
+#@permission_classes((AllowAny,))
+#def login(request):
+#    username = request.data.get("username")
+#    password = request.data.get("password")
+#    if username is None or password is None:
+#        return Response({'error': 'Please provide both username and password'},
+#                        status=HTTP_400_BAD_REQUEST)
+#    user = authenticate(username=username, password=password)
+#    if not user:
+#        return Response({'error': 'Invalid Credentials'},
+#                        status=HTTP_404_NOT_FOUND)
+#    token, _ = Token.objects.get_or_create(user=user)
+#    return Response({'token': token.key},
+#                    status=HTTP_200_OK)
 
     # Create your views here.
 
@@ -105,6 +132,12 @@ def listAchat(request):
     return Response(serialization.data)
 
 @api_view(['GET'])
+def achlist(request):
+    ligneAch = Panier.objects.prefetch_related('id')
+    serialization = lignAchatSerializer(ligneAch,many=True)
+    return Response(serialization.data)
+
+@api_view(['GET'])
 def allClientPanier(request):
     paniers = Panier.objects.all()
     serialization = panierSerializer(paniers,many=True)
@@ -118,10 +151,18 @@ def allClientPanier(request):
 
 # -------------------------------
 @api_view(['GET'])
-def getPanier(request,client_id):
+def getPanier(request,id):
+    panier = Panier.objects.get(id=id)
+    serializer = panierSerializer(panier)
+    return Response(serializer.data)
+
+# -------------------------------
+@api_view(['GET'])
+def getPanier1(request,client_id):
     panier = Panier.objects.filter(client__id=client_id)
     serializer = panierSerializer(panier)
     return Response(serializer.data)
+
 
 # -------------------------------
 @api_view(['POST'])
@@ -144,8 +185,11 @@ def updPanier(request,id):
 @api_view(['DELETE'])
 def delPanier(request,id):
     panier = Panier.objects.get(id=id)
-    panier.delete()
-    return Response('Suppression effectuée avec succès !!')
+    if panier.statut==True:
+        return Response('Impossible de supprimer un achat valide')
+    else:
+        panier.delete()
+        return Response('Suppression effectuée avec succès !!')
 
 #-----------------------------------#
 #     API GESTION DE LIGNE ACHAT    #
@@ -154,12 +198,6 @@ def delPanier(request,id):
 def allLigneAchat(request,panier_id):
     ligneAchats = LigneAchat.objects.filter(panier__id=panier_id)
     serialization = lignAchatSerializer(ligneAchats,many=True)
-    return Response(serialization.data)
-
-@api_view(['GET'])
-def achlist(request):
-    ligneAch = Panier.objects.prefetch_related('entetevte')
-    serialization = lignAchatSerializer(ligneAch,many=True)
     return Response(serialization.data)
 
 
@@ -193,3 +231,5 @@ def delLigneAchat(request,id):
     panier = Panier.objects.get(id=id)
     panier.delete()
     return Response('Suppression effectuée avec succès !!')
+
+
